pyproject.toml
[tool.poetry]
name = "semantic_testing"
version = "1.0.0"
description = "An application for running semantic tests on DITA content"
license = "Proprietary"
authors = ["Guidewire <xxx@xxx.xx>"]
readme = "README.md"
repository = "ssh://git@stash.guidewire.com/doctools/semantic-testing.git"
classifiers=[
    "Programming Language :: Python :: 3",
    "License :: Proprietary",
    "Operating System :: OS Independent",
]

[tool.poetry.dependencies]
python = "^3.8"
pyspellchecker = "^0.5.5"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = {version = "^20.8b1", allow-prereleases = true}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
